+++ include/SI/detail/detail.h
@@ -1,5 +1,5 @@
 /**
- * This file is part of "SI" version 2.0.0
+ * This file is part of "SI" version 2.0.1
  * A header only c++ library that provides type safety and user defined literals
  * for handling pyhsical values defined in the International System of
  * Units
@@ -28,8 +28,11 @@ template <typename _type> struct is_ratio : std::false_type {};
 template <intmax_t _num, intmax_t _den>
 struct is_ratio<std::ratio<_num, _den>> : std::true_type {};
 
-template <typename T, typename std::enable_if<
-                          std::is_floating_point<T>::value>::type * = nullptr>
+template <typename _type>
+inline constexpr bool is_ratio_v = is_ratio<_type>::value;
+
+template <typename T,
+          std::enable_if_t<std::is_floating_point_v<T>> * = nullptr>
 constexpr bool epsEqual(const T &lhs, const T &rhs) {
 
   return (lhs - rhs) < std::numeric_limits<T>::epsilon() &&