+++ include/SI/detail/unit.h
@@ -1,5 +1,5 @@
 /**
- * This file is part of "SI" version 2.0.0
+ * This file is part of "SI" version 2.0.1
  * A header only c++ library that provides type safety and user defined literals
  * for handling pyhsical values defined in the International System of
  * Units
@@ -51,11 +51,11 @@ template <typename _unit_lhs, typename _unit_rhs> struct unit_with_common_ratio;
 template <char _symbol, typename _exponent, typename _type,
           typename _ratio = std::ratio<1>>
 struct unit_t {
-  static_assert(std::is_arithmetic<_type>::value,
+  static_assert(std::is_arithmetic_v<_type>,
                 "Type is an arithmetic value");
-  static_assert(detail::is_ratio<_exponent>::value,
+  static_assert(detail::is_ratio_v<_exponent>,
                 "_exponent is a ratio type");
-  static_assert(detail::is_ratio<_ratio>::value, "_ratio is a std::ratio");
+  static_assert(detail::is_ratio_v<_ratio>, "_ratio is a std::ratio");
   using ratio = _ratio;
   using internal_type = _type;
   using exponent = _exponent;
@@ -81,11 +81,11 @@ struct unit_t {
   constexpr unit_t(const unit_t<_symbol, _exponent, _rhs_type, _rhs_ratio> &rhs)
       : value_{
             unit_cast<unit_t<_symbol, _exponent, _type, _ratio>>(rhs).value()} {
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before assigning");
   }
 
@@ -94,11 +94,11 @@ struct unit_t {
       : value_{
             std::move(unit_cast<unit_t<_symbol, _exponent, _type, _ratio>>(rhs)
                           .value())} {
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before assigning");
   }
 
@@ -116,16 +116,16 @@ struct unit_t {
 
   /// Assignment of same unit but different ratio
   template <typename _rhs_ratio,
-            typename std::enable_if<
-                !std::ratio_equal<_rhs_ratio, _ratio>::value>::type * = nullptr>
+            std::enable_if_t<!std::ratio_equal_v<_rhs_ratio, _ratio>> * =
+                nullptr>
   constexpr unit_t &
   operator=(const unit_t<_symbol, _exponent, _type, _rhs_ratio> &rhs) {
 
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before assigning");
 
     *this = unit_cast<unit_t<_symbol, _exponent, _type, _ratio>>(rhs);
@@ -134,16 +134,16 @@ struct unit_t {
 
   /// Move assignment of same unit but different ratio
   template <typename _rhs_ratio,
-            typename std::enable_if<
-                !std::ratio_equal<_rhs_ratio, _ratio>::value>::type * = nullptr>
+            std::enable_if_t<!std::ratio_equal_v<_rhs_ratio, _ratio>> * =
+                nullptr>
   constexpr unit_t &
   operator=(unit_t<_symbol, _exponent, _type, _rhs_ratio> &&rhs) {
 
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before assigning");
 
     *this =
@@ -159,19 +159,19 @@ struct unit_t {
 
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before comparing");
 
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
-    static_assert(std::is_integral<_type>::value ||
-                      std::is_floating_point<_type>::value,
+    static_assert(std::is_integral_v<_type> ||
+                      std::is_floating_point_v<_type>,
                   "Is integral or floating point");
     using gcd_unit = typename unit_with_common_ratio<
         typename std::remove_reference<decltype(rhs)>::type,
         typename std::remove_reference<decltype(*this)>::type>::type;
 
-    if constexpr (std::is_integral<_type>::value) {
+    if constexpr (std::is_integral_v<_type>) {
 
       return unit_cast<gcd_unit>(rhs).value() ==
              unit_cast<gcd_unit>(*this).value();
@@ -185,7 +185,7 @@ struct unit_t {
   template <typename _rhs_type, typename _rhs_ratio>
   constexpr bool operator!=(
       const unit_t<_symbol, _exponent, _rhs_type, _rhs_ratio> &rhs) const {
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     return !(*this == rhs);
   }
@@ -193,11 +193,11 @@ struct unit_t {
   template <typename _rhs_type, typename _rhs_ratio>
   constexpr bool operator<(
       const unit_t<_symbol, _exponent, _rhs_type, _rhs_ratio> &rhs) const {
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before comparing");
 
     using gcd_unit = typename unit_with_common_ratio<
@@ -216,11 +216,11 @@ struct unit_t {
   template <typename _rhs_type, typename _rhs_ratio>
   constexpr bool operator>(
       const unit_t<_symbol, _exponent, _rhs_type, _rhs_ratio> &rhs) const {
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before comparing");
 
     using gcd_unit = typename unit_with_common_ratio<
@@ -245,7 +245,7 @@ struct unit_t {
   constexpr auto operator*(
       const unit_t<_symbol, _rhs_exponent, _rhs_type, _ratio> &rhs) const {
 
-    static_assert(detail::is_ratio<_rhs_exponent>::value,
+    static_assert(detail::is_ratio_v<_rhs_exponent>,
                   "rhs exponent is a ratio type");
     return unit_t<_symbol, std::ratio_add<_rhs_exponent, _exponent>, _type,
                   std::ratio_multiply<ratio, _ratio>>{value() * rhs.value()};
@@ -258,14 +258,14 @@ struct unit_t {
   template <typename _rhs_exponent, typename _rhs_ratio, typename _rhs_type>
   constexpr auto operator*(
       const unit_t<_symbol, _rhs_exponent, _rhs_type, _rhs_ratio> &rhs) const {
-    static_assert(detail::is_ratio<_rhs_exponent>::value,
+    static_assert(detail::is_ratio_v<_rhs_exponent>,
                   "rhs exponent is a ratio type");
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
 
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before comparing");
 
     return unit_t<_symbol, std::ratio_add<_exponent, _rhs_exponent>, _type,
@@ -283,12 +283,13 @@ struct unit_t {
 
   /// divide with same unit with same ratio but not the same exponent
   /// @returns unit with exponents subtracted from each others
-  template <typename _rhs_exponent, typename _rhs_type,
-            typename std::enable_if<std::ratio_not_equal<
-                _rhs_exponent, _exponent>::value>::type * = nullptr>
+  template <
+      typename _rhs_exponent, typename _rhs_type,
+      std::enable_if_t<std::ratio_not_equal_v<_rhs_exponent, _exponent>>
+          * = nullptr>
   constexpr auto operator/(
       const unit_t<_symbol, _rhs_exponent, _rhs_type, _ratio> &rhs) const {
-    static_assert(detail::is_ratio<_rhs_exponent>::value,
+    static_assert(detail::is_ratio_v<_rhs_exponent>,
                   "rhs exponent is a ratio type");
     using rhs_t = typename std::remove_reference<decltype(rhs)>::type;
 
@@ -301,18 +302,19 @@ struct unit_t {
   /// divide with a same unit but different ratios
   /// the ratio of the result is the gcd of the two ratios and the exponents are
   /// subtracted
-  template <typename _rhs_exponent, typename _rhs_type, typename _rhs_ratio,
-            typename std::enable_if<std::ratio_not_equal<
-                _rhs_exponent, _exponent>::value>::type * = nullptr>
+  template <
+      typename _rhs_exponent, typename _rhs_type, typename _rhs_ratio,
+      std::enable_if_t<std::ratio_not_equal_v<_rhs_exponent, _exponent>>
+          * = nullptr>
   constexpr auto operator/(
       const unit_t<_symbol, _rhs_exponent, _rhs_type, _rhs_ratio> &rhs) const {
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
-    static_assert(detail::is_ratio<_rhs_exponent>::value,
+    static_assert(detail::is_ratio_v<_rhs_exponent>,
                   "rhs exponent is a ratio type");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before dividing");
 
     return unit_t<_symbol, std::ratio_subtract<exponent, _rhs_exponent>, _type,
@@ -329,19 +331,19 @@ struct unit_t {
   /// if the same units of the same exponent but different ratio are divided
   /// then the result is a scalar
   template <typename _rhs_exponent, typename _rhs_type, typename _rhs_ratio,
-            typename std::enable_if<std::ratio_equal<
-                _rhs_exponent, exponent>::value>::type * = nullptr>
+            std::enable_if_t<std::ratio_equal_v<_rhs_exponent, exponent>>
+                * = nullptr>
   constexpr _type operator/(
       const unit_t<_symbol, _rhs_exponent, _rhs_type, _rhs_ratio> &rhs) const {
     static_assert(SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-                      std::ratio_equal<_rhs_ratio, _ratio>::value,
+                      std::ratio_equal_v<_rhs_ratio, _ratio>,
                   "Implicit ratio conversion disabled, convert to same ratio "
                   "before dividing");
 
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
 
-    static_assert(detail::is_ratio<_rhs_exponent>::value,
+    static_assert(detail::is_ratio_v<_rhs_exponent>,
                   "rhs exponent is a ratio type");
 
     using gcd_unit = typename unit_with_common_ratio<
@@ -362,11 +364,11 @@ struct unit_t {
   constexpr unit_t operator+(
       const unit_t<_symbol, _exponent, _rhs_type, _rhs_ratio> &rhs) const {
 
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before adding values");
 
     return unit_t{
@@ -382,16 +384,16 @@ struct unit_t {
 
   /// add value of the same type but possibly different ratio
   template <typename _rhs_type, typename _rhs_ratio,
-            typename std::enable_if<
-                !std::ratio_equal<_rhs_ratio, _ratio>::value>::type * = nullptr>
+            std::enable_if_t<!std::ratio_equal_v<_rhs_ratio, _ratio>> * =
+                nullptr>
   constexpr unit_t &
   operator+=(const unit_t<_symbol, _exponent, _rhs_type, _rhs_ratio> &rhs) {
 
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before adding values");
 
     value_ += unit_cast<unit_t<_symbol, _exponent, _type, _ratio>>(rhs).value();
@@ -404,11 +406,11 @@ struct unit_t {
   constexpr unit_t operator-(
       const unit_t<_symbol, _exponent, _rhs_type, _rhs_ratio> &rhs) const {
 
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before subtracting");
 
     return unit_t{
@@ -423,17 +425,17 @@ struct unit_t {
   }
 
   /// subtract value of the same type but possibly different ratio
-  template <typename _rhs_type, typename _rhs_ratio,
-            typename std::enable_if<
-                !std::ratio_equal<_rhs_ratio, _ratio>::value>::type * = nullptr>
+  template <
+      typename _rhs_type, typename _rhs_ratio,
+      std::enable_if<!std::ratio_equal_v<_rhs_ratio, _ratio>> * = nullptr>
   constexpr unit_t &
   operator-=(const unit_t<_symbol, _exponent, _type, _rhs_ratio> &rhs) {
 
-    static_assert(detail::is_ratio<_rhs_ratio>::value,
+    static_assert(detail::is_ratio_v<_rhs_ratio>,
                   "_rhs_ratio is a std::ratio");
     static_assert(
         SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-            std::ratio_equal<ratio, _rhs_ratio>::value,
+            std::ratio_equal_v<ratio, _rhs_ratio>,
         "Implicit ratio conversion disabled, convert before adding values");
 
     value_ -= unit_cast<unit_t<_symbol, _exponent, _type, _ratio>>(rhs).value();
@@ -478,10 +480,9 @@ private:
 /// operator to divide primitive type by unit encapsulating the same type
 /// template specialisation handling integer types
 /// @results unit with negative exponent
-template <
-    typename _type, char _symbol, typename _exponent, typename _rhs_type,
-    typename _ratio,
-    typename std::enable_if<std::is_integral<_type>::value>::type * = nullptr>
+template <typename _type, char _symbol, typename _exponent, typename _rhs_type,
+          typename _ratio,
+          std::enable_if_t<std::is_integral_v<_type>> * = nullptr>
 constexpr auto
 operator/(const _type &lhs,
           const unit_t<_symbol, _exponent, _rhs_type, _ratio> &rhs) {
@@ -489,7 +490,7 @@ operator/(const _type &lhs,
                     std::ratio_equal<std::ratio<1>, _ratio>::value,
                 "Implicit ratio conversion disabled, convert to ratio<1> "
                 "before dividing");
-  static_assert(detail::is_ratio<_exponent>::value, "Exponent is a ratio type");
+  static_assert(detail::is_ratio_v<_exponent>, "Exponent is a ratio type");
   return unit_cast<unit_t<
       _symbol, std::ratio_multiply<std::ratio<-1>, _exponent>, _type, _ratio>>(
       unit_t<_symbol, std::ratio_multiply<std::ratio<-1>, _exponent>, _type,
@@ -502,16 +503,15 @@ operator/(const _type &lhs,
 /// @results unit with negative exponent
 template <typename _type, char _symbol, typename _exponent, typename _rhs_type,
           typename _ratio,
-          typename std::enable_if<std::is_floating_point<_type>::value>::type
-              * = nullptr>
+          std::enable_if_t<std::is_floating_point_v<_type>> * = nullptr>
 constexpr auto
 operator/(const _type &lhs,
           const unit_t<_symbol, _exponent, _rhs_type, _ratio> &rhs) {
   static_assert(SI_ENABLE_IMPLICIT_RATIO_CONVERSION ||
-                    std::ratio_equal<_ratio, std::ratio<1>>::value,
+                    std::ratio_equal_v<_ratio, std::ratio<1>>,
                 "Implicit ratio conversion disabled, convert to ratio<1> "
                 "before dividing");
-  static_assert(detail::is_ratio<_exponent>::value, "Exponent is a ratio type");
+  static_assert(detail::is_ratio_v<_exponent>, "Exponent is a ratio type");
   return unit_cast<unit_t<
       _symbol, std::ratio_multiply<std::ratio<-1>, _exponent>, _type, _ratio>>(
       unit_t<_symbol, std::ratio_multiply<std::ratio<-1>, _exponent>, _type,
@@ -532,13 +532,16 @@ struct is_unit_t<const unit_t<_symbol, _exponent, _type, _ratio>>
 template <char _symbol, typename _exponent, typename _ratio, typename _type>
 struct is_unit_t<unit_t<_symbol, _exponent, _type, _ratio>> : std::true_type {};
 
+template <typename _type>
+inline constexpr bool is_unit_t_v = is_unit_t<_type>::value;
+
 /// function to cast between two units of the same type
 template <typename _target_type, typename _rhs_T>
 constexpr auto unit_cast(const _rhs_T &rhs) {
   // using static assert instead of std::enable if in order to be able to
   // forward declare this function easier
   static_assert(
-      is_unit_t<_rhs_T>::value ||
+      is_unit_t_v<_rhs_T> ||
           std::is_base_of<
               unit_t<_rhs_T::symbol::value, typename _rhs_T::exponent,
                      typename _rhs_T::internal_type, typename _rhs_T::ratio>,
@@ -553,8 +556,8 @@ constexpr auto unit_cast(const _rhs_T &rhs) {
 
 template <typename _unit_lhs, typename _unit_rhs>
 struct unit_with_common_ratio {
-  static_assert(is_unit_t<_unit_lhs>::value, "only supported for SI::unit_t");
-  static_assert(is_unit_t<_unit_rhs>::value, "only supported for SI::unit_t");
+  static_assert(is_unit_t_v<_unit_lhs>, "only supported for SI::unit_t");
+  static_assert(is_unit_t_v<_unit_rhs>, "only supported for SI::unit_t");
   static_assert(std::is_convertible<typename _unit_lhs::internal_type,
                                     typename _unit_rhs::internal_type>::value);
   static_assert(_unit_lhs::symbol::value == _unit_rhs::symbol::value);
@@ -575,8 +578,8 @@ constexpr auto cross_unit_divide(const _unit_lhs &lhs, const _unit_rhs &rhs) {
   // do not use for the same unit as this should result in decreasing the
   // exponent
   static_assert(!std::is_same<_unit_lhs, _unit_rhs>::value);
-  static_assert(is_unit_t<_unit_lhs>::value, "lhs parameter is a unit_t");
-  static_assert(is_unit_t<_unit_rhs>::value, "rhs parameter is a unit_t");
+  static_assert(is_unit_t_v<_unit_lhs>, "lhs parameter is a unit_t");
+  static_assert(is_unit_t_v<_unit_rhs>, "rhs parameter is a unit_t");
 
   using resulting_ratio = typename std::ratio_divide<typename _unit_lhs::ratio,
                                                      typename _unit_rhs::ratio>;
@@ -595,8 +598,8 @@ constexpr auto cross_unit_multiply(const _unit_lhs &lhs, const _unit_rhs &rhs) {
   // do not use for the same unit as this should result in increasing the
   // exponent
   static_assert(!std::is_same<_unit_lhs, _unit_rhs>::value);
-  static_assert(is_unit_t<_unit_lhs>::value, "lhs parameter is a unit_t");
-  static_assert(is_unit_t<_unit_rhs>::value, "rhs parameter is a unit_t");
+  static_assert(is_unit_t_v<_unit_lhs>, "lhs parameter is a unit_t");
+  static_assert(is_unit_t_v<_unit_rhs>, "rhs parameter is a unit_t");
   using resulting_ratio =
       typename std::ratio_multiply<typename _unit_lhs::ratio,
                                    typename _unit_rhs::ratio>;